+++ src/main/java/com/example/buildingmanager/models/admin/UpdateAndCreateBuildingDTO.java
@@ -24,7 +24,7 @@ public class UpdateAndCreateBuildingDTO {
     private String direction;
     private String level;
 
-    private Integer rentPrice;
+    private Long rentPrice;
     private String rentPriceDescription;
     private String serviceFee;
     private String carFee;
